1-3_7-3/requirements.txt
openai==2.9.0
google-generativeai==0.8.5
python-dotenv==1.2.1
